feat: modifications on services layer

+++ src/main/java/com/ricael/mergemind/services/ApplicationServices.java
@@ -1,7 +1,11 @@
 package com.ricael.mergemind.services;
 
 import com.ricael.mergemind.domain.Application;
+import com.ricael.mergemind.domain.Role;
+import com.ricael.mergemind.domain.User;
+import com.ricael.mergemind.domain.enums.Status;
 import com.ricael.mergemind.dto.mapper.ApplicationMapper;
+import com.ricael.mergemind.dto.mapper.RoleMapper;
 import com.ricael.mergemind.dto.request.ApplicationRequest;
 import com.ricael.mergemind.dto.response.ApplicationResponse;
 import com.ricael.mergemind.repository.ApplicationRepository;
@@ -16,8 +20,22 @@ public class ApplicationServices {
     @Autowired
     ApplicationRepository applicationRepository;
 
+    @Autowired
+    UserServices userServices;
+
+    @Autowired
+    RoleServices roleServices;
+
     public ApplicationResponse create(ApplicationRequest applicationRequest) {
-        return ApplicationMapper.toResponse(applicationRepository.save(ApplicationMapper.toEntity(applicationRequest)));
+        Role r = roleServices.getRoleById(applicationRequest.role().id());
+        User user = userServices.getUserEntityById(applicationRequest.user().id());
+
+        Application app = new Application();
+        app.setStatus(applicationRequest.status());
+        app.setRole(r);
+        app.setUser(user);
+
+        return ApplicationMapper.toResponse(applicationRepository.save(app));
     }
 
     public List<ApplicationResponse> findByUserId(Long userId) {
@@ -34,12 +52,21 @@ public class ApplicationServices {
                 .toList();
     }
 
-    public ApplicationResponse updateApplicationStatus(Long id, String status) {
-        Application updatedApplication = applicationRepository.updateApplicationStatus(id, status);
-        return ApplicationMapper.toResponse(updatedApplication);
+    public void updateApplicationStatus(Long id, String status) {
+
+        try{
+            Status statusEnum = Status.valueOf(status.toUpperCase());
+            applicationRepository.updateApplicationStatus(id, statusEnum);
+        } catch (IllegalArgumentException e) {
+            throw new IllegalArgumentException("Invalid status value: " + status);
+        }
+
     }
 
     public void deleteApplication(Long id) {
+        if(!applicationRepository.existsById(id)) {
+            throw new IllegalArgumentException("Application with id " + id + " does not exist.");
+        }
         applicationRepository.deleteById(id);
     }
 }
